Refactor study mongo
